Updates to dht11: humidity int to char; Updates to slave and master node: main_noed.cpp->Line 23: add sensor code

+++ dht11.h
@@ -36,14 +36,14 @@ public:
      * @returns
      * Celsius int
      */
-     int getCelsius();
+     uint8_t getCelsius();
 
     /** Get the humidity from the saved object.
      *
      * @returns
      * Humidity percentage int
      */
-     int getHumidity();
+     uint8_t getHumidity();
 
 private:
   ///percentage of humidity
